+++ sdk/keyvault/azure-security-keyvault-keys/CMakeLists.txt
@@ -30,7 +30,6 @@ set(
   AZURE_KEYVAULT_KEYS_HEADER
     inc/azure/keyvault/keys/cryptography/cryptography_client_options.hpp
     inc/azure/keyvault/keys/cryptography/cryptography_client.hpp
-    inc/azure/keyvault/keys/cryptography/cryptography_provider.hpp
     inc/azure/keyvault/keys/cryptography/cryptography_serializers.hpp
     inc/azure/keyvault/keys/cryptography/decrypt_parameters.hpp
     inc/azure/keyvault/keys/cryptography/decrypt_result.hpp
@@ -41,18 +40,18 @@ set(
     inc/azure/keyvault/keys/cryptography/key_verify_parameters.hpp
     inc/azure/keyvault/keys/cryptography/key_wrap_algorithm.hpp
     inc/azure/keyvault/keys/cryptography/key_wrap_parameters.hpp
-    inc/azure/keyvault/keys/cryptography/local_cryptography_provider_factory.hpp
-    inc/azure/keyvault/keys/cryptography/local_cryptography_provider.hpp
-    inc/azure/keyvault/keys/cryptography/remote_cryptography_client.hpp
-    inc/azure/keyvault/keys/cryptography/rsa_cryptography_provider.hpp
     inc/azure/keyvault/keys/cryptography/sign_result.hpp
     inc/azure/keyvault/keys/cryptography/signature_algorithm.hpp
     inc/azure/keyvault/keys/cryptography/wrap_result.hpp
     inc/azure/keyvault/keys/cryptography/unwrap_result.hpp
     inc/azure/keyvault/keys/cryptography/verify_result.hpp
+    inc/azure/keyvault/keys/internal/cryptography/cryptography_provider.hpp
+    inc/azure/keyvault/keys/internal/cryptography/local_cryptography_provider_factory.hpp
+    inc/azure/keyvault/keys/internal/cryptography/local_cryptography_provider.hpp
+    inc/azure/keyvault/keys/internal/cryptography/remote_cryptography_client.hpp
+    inc/azure/keyvault/keys/internal/cryptography/rsa_cryptography_provider.hpp
     inc/azure/keyvault/keys/delete_key_operation.hpp
     inc/azure/keyvault/keys/deleted_key.hpp
-    inc/azure/keyvault/keys/details/key_request_parameters.hpp
     inc/azure/keyvault/keys/import_key_options.hpp
     inc/azure/keyvault/keys/json_web_key.hpp
     inc/azure/keyvault/keys/key_client.hpp
@@ -89,6 +88,8 @@ set(
     src/cryptography/verify_result.cpp
     src/private/key_backup.hpp
     src/private/key_constants.hpp
+    src/private/key_request_parameters.hpp
+    src/private/key_serializers.hpp
     src/private/package_version.hpp
     src/delete_key_operation.cpp
     src/deleted_key.cpp
