+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/internal/cryptography/cryptography_provider.hpp
