+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/internal/cryptography/rsa_cryptography_provider.hpp
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "azure/keyvault/keys/cryptography/local_cryptography_provider.hpp"
+#include "azure/keyvault/keys/internal/cryptography/local_cryptography_provider.hpp"
 
 #include <memory>
 #include <string>
