+++ sdk/keyvault/azure-security-keyvault-keys/src/private/key_serializers.hpp
